Update applications-that-can-bypass-appcontrol.md
I am on of the SPMs with MSRC and the case was assessed as Moderate severity and the action was to fix online content of the blog. 

Here is the feedback received from the engineering team.

In this link, if you look at the following value

<Deny ID="ID_DENY_ADDINPROCESS32" FriendlyName="AddInProcess32.exe" FileName="AddInProcess32.exe" MinimumFileVersion="0.0.0.0" MaximumFileVersion="65355.65355.65355.65355" />

The MaximumFileVersion value is 65355 which should be corrected to 65535. This is an example and it should be corrected for all the binaries.

referene:
https://learn.microsoft.com/en-us/visualstudio/deployment/how-to-specify-where-visual-studio-copies-the-files?view=vs-2019#:~:text=maximum%20value%20of-,65535,-(MaxValue)%2C%20the